requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Background Tasks & Caching
celery==5.3.4
redis==5.0.1
flower==2.0.1

# External APIs
google-cloud-vision==3.4.5
plaid-python==11.0.0
twilio==8.10.3
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Data Processing & Validation
pydantic[email]==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0
python-dateutil==2.8.2
Pillow==10.1.0

# HTTP & Utilities
httpx==0.25.2
aiofiles==23.2.1
python-dotenv==1.0.0

# Monitoring & Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Production
prometheus-client==0.19.0

# System Monitoring
psutil==5.9.6

# Additional Security
cryptography==41.0.7
